src/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
django-multiselectfield==1.0.1
psycopg==3.2.12
sqlparse==0.5.3
